+++ src/Client/Contracts/Client.php
@@ -25,14 +25,14 @@ interface Client
     /**
      * Version of YouTrack REST PHP client.
      */
-    const VERSION = '3.1.1';
+    const VERSION = '3.2.0';
 
     /**
      * Create and send an HTTP request.
      *
      * @param string $method
      * @param string $uri
-     * @param array $formData
+     * @param array $params
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -40,13 +40,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function request(string $method, string $uri, array $formData = [], array $options = []): ResponseContract;
+    public function request(string $method, string $uri, array $params = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an GET HTTP request.
      *
      * @param string $uri
-     * @param array $formData
+     * @param array $params
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -54,13 +54,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function get(string $uri, array $formData = [], array $options = []): ResponseContract;
+    public function get(string $uri, array $params = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an POST HTTP request.
      *
      * @param string $uri
-     * @param array $formData
+     * @param array $params
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -68,13 +68,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function post(string $uri, array $formData = [], array $options = []): ResponseContract;
+    public function post(string $uri, array $params = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an PUT HTTP request.
      *
      * @param string $uri
-     * @param array $formData
+     * @param array $params
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -82,13 +82,13 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function put(string $uri, array $formData = [], array $options = []): ResponseContract;
+    public function put(string $uri, array $params = [], array $options = []): ResponseContract;
 
     /**
      * Create and send an DELETE HTTP request.
      *
      * @param string $uri
-     * @param array $formData
+     * @param array $params
      * @param array $options
      * @return \Cog\YouTrack\Rest\Response\Contracts\Response
      *
@@ -96,13 +96,34 @@ interface Client
      * @throws \Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException
      * @throws \Cog\YouTrack\Rest\Client\Exceptions\ClientException
      */
-    public function delete(string $uri, array $formData = [], array $options = []): ResponseContract;
+    public function delete(string $uri, array $params = [], array $options = []): ResponseContract;
 
     /**
      * Write header value.
      *
      * @param string $key
      * @param string $value
+     * @return void
+     */
+    public function withHeader(string $key, string $value): void;
+
+    /**
+     * Write header values.
+     *
+     * @param array $headers
+     * @return void
+     */
+    public function withHeaders(array $headers): void;
+
+    /**
+     * Write header value.
+     *
+     * @deprecated 3.2.0
+     * @see withHeader
+     *
+     * @param string $key
+     * @param string $value
+     * @return void
      */
     public function putHeader(string $key, string $value): void;
 }